+++ packages/router/src/router.tsx
@@ -65,6 +65,7 @@ interface Config {
 export interface NavigationOptions {
 	transition?: string;
 	state?: Record< string, any >;
+	replace?: boolean;
 }
 
 const RoutesContext = createContext< Match | null >( null );
@@ -106,7 +107,7 @@ export function useHistory() {
 				const result = beforeNavigate
 					? beforeNavigate( { path, query } )
 					: { path, query };
-				return history.push(
+				return history[ options.replace ? 'replace' : 'push' ](
 					{
 						search: buildQueryString( {
 							[ pathArg ]: result.path,