+++ packages/router/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   `useHistory().navigate` accepts a `replace` option to replace the current history entry instead of pushing a new one.
+
 ## 1.50.0 (2026-07-01)
 
 ## 1.49.0 (2026-06-24)